vi-feat: store reducers placed in slices folder && reducers replaced with slices
