--- client/src/packages/store/actions.js
@@ -1,4 +1,0 @@
-export * as appActionCreator from './app/actions.js';
-export * as profileActionCreator from './profile/actions.js';
-export * as threadActionCreator from './thread/actions.js';
-export * as notificationActionCreator from './notifications/actions.js';
