--- client/src/packages/store/root-reducer.js
@@ -1,2 +1,0 @@
-export { reducer as profileReducer } from './profile/reducer.js';
-export { reducer as threadReducer } from './thread/reducer.js';
