+++ client/src/packages/store/store.js
@@ -7,7 +7,8 @@ import { comment } from 'packages/comment/comment.js';
 import { post } from 'packages/post/post.js';
 import { image } from 'packages/image/image.js';
 import { notification } from 'packages/notification/notification.js';
-import { profileReducer, threadReducer } from './root-reducer.js';
+import { reducer as threadReducer } from 'slices/thread/thread.js';
+import { reducer as profileReducer } from 'slices/profile/profile.js';
 import { notificationSocket } from './middlewares/middlewares.js';
 
 const store = configureStore({
