+++ client/src/pages/sign/sign.jsx
@@ -1,6 +1,6 @@
 import { useCallback, useDispatch, useLocation } from 'libs/hooks/hooks.js';
 import { AppRoute } from 'libs/enums/enums.js';
-import { profileActionCreator } from 'packages/store/actions.js';
+import { actions as profileActionCreator } from 'slices/profile/profile.js';
 import { Image } from 'libs/components/image/image.jsx';
 import { LoginForm, RegistrationForm } from './components/components.js';
 import styles from './styles.module.scss';
