+++ client/src/slices/app/app.js
@@ -1,0 +1,7 @@
+import { notify } from './actions.js';
+
+const actions = {
+  notify
+};
+
+export { actions };
