+++ client/src/slices/notifications/notifications.js
@@ -1,0 +1,8 @@
+import { joinRoom, leaveRoom } from './actions.js';
+
+const actions = {
+  joinRoom,
+  leaveRoom
+};
+
+export { actions };
